+++ test/util/mvc/controllers/gracefulShutdownController.js
@@ -1,6 +1,5 @@
 module.exports = router => {
   router.route('/longWait').get((req, res) => {
-    process.send('waiting')
     setTimeout(() => {
       res.send('longWait done')
     }, 1000)